Move IndexedView in its own file
