+++ collections/src/main/scala/strawman/collection/View.scala
@@ -294,70 +294,3 @@ object View extends IterableFactory[View] {
     override def knownSize: Int = if (underlying.knownSize >= 0) underlying.knownSize max len else -1
   }
 }
-
-/** View defined in terms of indexing a range */
-trait IndexedView[+A] extends View[A] with SeqOps[A, View, View[A]] { self =>
-
-  def iterator(): Iterator[A] = new AbstractIterator[A] {
-    private var current = 0
-    override def knownSize: Int = self.length - current
-    def hasNext = current < self.length
-    def next(): A = {
-      val r = self.apply(current)
-      current += 1
-      r
-    }
-  }
-
-  final override def knownSize: Int = length
-
-  override def take(n: Int): IndexedView[A] = new IndexedView.Take(this, n)
-  override def takeRight(n: Int): IndexedView[A] = new IndexedView.TakeRight(this, n)
-  override def drop(n: Int): IndexedView[A] = new IndexedView.Drop(this, n)
-  override def dropRight(n: Int): IndexedView[A] = new IndexedView.DropRight(this, n)
-  override def map[B](f: A => B): IndexedView[B] = new IndexedView.Map(this, f)
-  override def reverse: IndexedView[A] = IndexedView.Reverse(this)
-}
-
-object IndexedView {
-
-  class Take[A](underlying: IndexedView[A], n: Int) extends IndexedView[A] {
-    private[this] val normN = n max 0
-    def length = underlying.length min normN
-    @throws[IndexOutOfBoundsException]
-    def apply(i: Int) = underlying.apply(i)
-  }
-
-  class TakeRight[A](underlying: IndexedView[A], n: Int) extends IndexedView[A] {
-    private[this] val delta = (underlying.length - (n max 0)) max 0
-    def length = underlying.length - delta
-    @throws[IndexOutOfBoundsException]
-    def apply(i: Int) = underlying.apply(i + delta)
-  }
-
-  class Drop[A](underlying: IndexedView[A], n: Int) extends IndexedView[A] {
-    protected val normN = n max 0
-    def length = (underlying.length - normN) max 0
-    @throws[IndexOutOfBoundsException]
-    def apply(i: Int) = underlying.apply(i + normN)
-  }
-
-  class DropRight[A](underlying: IndexedView[A], n: Int) extends IndexedView[A] {
-    private[this] val len = (underlying.length - (n max 0)) max 0
-    def length = len
-    @throws[IndexOutOfBoundsException]
-    def apply(i: Int) = underlying.apply(i)
-  }
-
-  class Map[A, B](underlying: IndexedView[A], f: A => B) extends IndexedView[B] {
-    def length = underlying.length
-    @throws[IndexOutOfBoundsException]
-    def apply(n: Int) = f(underlying.apply(n))
-  }
-
-  case class Reverse[A](underlying: IndexedView[A]) extends IndexedView[A] {
-    def length = underlying.length
-    @throws[IndexOutOfBoundsException]
-    def apply(i: Int) = underlying.apply(length - 1 - i)
-  }
-}
